test: Policies: Test fromCIDR properly
The original test here specified different containers for the
fromEndpoints and fromCIDRs sections of this policy, so the union of
these should allow both of those containers to access this endpoint (and
others would be locked out). However, this test is currently only
testing that the fromEndpoints policy is applied correctly.

Swap out the fromEndpoints to app2. I'd expect that app1 should still be
allowed based on fromCIDR, however this appears not to be the case.

needs-triage.

Signed-off-by: Joe Stringer <joe@covalent.io>