cilium: gingko option 'useVagrant' to run tests without VMs
Running tests on bare-metal is useful for doing performance analysis
and development work on the datapath in the Linux kernel. The old
test bash scripts could easily be run on bare-metal by simply calling
the script on the system under test.

However, the new gingko test framework insists on launching vagrant
boxes. To avoid this add a new 'useVagrant' flag. The following cmd
will allow running on bare metal,

 $ gingko --focus="Runtime*" -- \
         -cilium.provision=false -cilium.useVagrant=false

Not all tests currently pass with this option, but individual test
failures can be addressed in follow on patches. We may want to
consider adding a --focus class for bare metal testing.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>