cilium: Print cluster status in 'cilium status'
This commit adds a cluster status to `cilium status` via an API call to
`cilium-health`. By default, it prints a single line listing how many
nodes are reachable and the timestamp:

    $ cilium status
    ...
    Cluster health:   1/1 reachable   (2018-02-16T13:25:53-08:00)

If any nodes or health endpoints are unreachable, they will be listed
below this message (up to 10 entries by default).

In verbose mode, it also prints every node and its status:

    $ cilium status --all-health
    Cluster health:   1/1 reachable   (2018-02-16T13:30:53-08:00)
      Name            IP              Reachable   Endpoints reachable
      cilium-master   10.0.2.15       true        false

Signed-off-by: Joe Stringer <joe@covalent.io>