packet: Default to x1.small.x86
The previous setting was somehow defaulting to the atom processors,
t1.small.x86, which only had 8GB of RAM, causing problems with the
default VM memory configuration of the VMs which lead to random
OOM-kills of the VMs while attempting to run ginkgo. x1.small.x86 is
almost 5x more expensive, but you'll save that much time having an
instance that actually works.

Signed-off-by: Joe Stringer <joe@cilium.io>