examples: Add proxy connectivity checks
These new checks configure various L7 proxy paths to validate
connectivity via L7 proxies, in the following dimensions:

- Allow / Deny
- Intra-node / Multi-node
- Apply policy on egress; ingress; or both (proxy-to-proxy)

Note that proxy-to-proxy always configures egress allow policy to ensure
that the traffic goes via the proxy and in the drop case the requests
are only rejected at the destination. This is because applying egress deny
at the source would prevent proxy-to-proxy connectivity.

These are implemented partially through a new 'echo-c' pod which always
has ingress policy applied to allow GET requests to '/'. Depending on
whether ingress policy is needed to check the particular permutation the
new checks may connect to 'echo-a' or 'echo-c'.

The 'connectivity-check-proxy.yaml' includes all of the above, plus the
existing FQDN test (which doesn't really have other permutations because
DNS policy is egress-only).

Finally, one dimension that's not covered here is IPv4 vs. IPv6. That
dimension is left up to the tester based on the cluster configuration.

Signed-off-by: Joe Stringer <joe@cilium.io>