identity, labels: Add kube-apiserver identity and label
This commit adds the new kube-apiserver label which enables traffic to
be identified as kube-apiserver traffic. Along with the new label, we
create a new reserved ID which will be associated with traffic going to
the kube-apiserver.

There are a couple nuances to note here. The kube-apiserver label will
used in three scenarios:
  * Attached along with the local host label if the kube-apiserver is
    running on the local host. The identity of the local host will now
    be used for traffic to the local host + to the local kube-apiserver.
    The reserved ID of the local host *does not* change however. It is
    still 1, to prevent breaking changes with connectivity.
  * Attached along with the remote-node label as part of the new
    kube-apiserver reserved identity. This identity is only used if the
    kube-apiserver is running on remote nodes.
  * Attached along with the CIDR labels if the kube-apiserver is running
    outside of the cluster. This will result in a regular CIDR identity.

In subsequent commits, the above three cases are what will be
implemented to allow policy matching on the kube-apiserver. This commit
setting the groundwork for that.

Signed-off-by: Chris Tarazi <chris@isovalent.com>