k8s/watchers: Push kube-apiserver IPs into IDMD
By leveraging the special "kubernetes" Service / Endpoint object, we can
learn about which IPs the kube-apiserver is behind. These IPs will be
upserted into the ipcache with an identity that is to-be-allocated.  In
order to allocate an identity for these IPs, the
kube-apiserver-associated IPs are inserted into the IDMD map with the
"reserved:kube-apiserver" label. Then the watcher will trigger the label
injection via ipcache.TriggerLabelInjection().

We do not make a distinction on where the kube-apiserver is running,
whether it's on the local node, on a remote node in the cluster, or if
it's running outside of the cluster. All kube-apiserver-associated IPs
will have the "reserved:kube-apiserver" label.

Note that we are not hooking into the delete event. We only subscribe to
the add / update events. This is because we perform a diff on the
currently associated kube-apiserver IPs inside the IDMD and what the add
/ update event contains (desired IPs), then remove the current IPs that
are not in the desired set. This operation is equivalent to subscribing
to the delete event and removing the "desired" IPs from the IDMD.
However, we don't need to subscribe to the delete event because this
special Service is automatically recreated / reconciled by the apiserver
itself if the object is removed or edited.

Signed-off-by: Chris Tarazi <chris@isovalent.com>