Move CRD identity allocation mode to NewDaemon
TODO: Explain the consequences here

Move the CRD identity allocation mode check from initEnv to NewDaemon as it
checks if k8s-client is enabled. Since k8s configuration will live in
client.Config in an upcoming commit, we cannot perform this check in
initEnv() anymore as that happens before client.Config is populated and
the k8s client is created.

Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>