daemon: Refactor and combine datapath initialization
Move the datapath initialization code to a new dedicated function. The
primary functional change here is to also refactor the node address
initialization to happen at the same time. In a future commit, this will
need to occur after the LocalNodeStore has been initialized, so in
preparation we will move it here and this will simplify later changes.

Signed-off-by: Jussi Maki <jussi@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>