Add option --empty-service-mode
Commit 183501124869 ("bpf: drop SVC traffic if no backend is available")
forced a drop for traffic towards services that have no backends
available. However, for some use cases it can be useful to pass this
traffic instead, primarily as a method to induce a proper TCP RST or
ICMP DEST_UNREACHABLE message back to the caller.

This commit implements a new --empty-service-mode flag in order to
expose this capability to the user, so they can decide how such traffic
should be handled.

This only applies to traffic from Pods, it does not apply to NodePort
traffic. Applying this technique for NodePort traffic could be applied
on top.

Signed-off-by: Joe Stringer <joe@cilium.io>