workflows/e2e: Apply connectivity test timeout at 15m
At the moment, the full CLI testsuite takes around 8 minutes. Rounding
to 10 and adding 5 more minutes brings the number up to 15m, so set this
as the CLI timeout.

The goal here is to ensure that in the event of a failure condition that
may lead to the CLI timing out (such as every test failing for the same
reason, or some fundamental issue in the infrastructure), the CLI will
terminate early, gather a sysdump, and return before GitHub timeouts
kick in. This should ensure that even in such adverse situations,
developers will have test results they can consult to identify what went
wrong in a workflow.

Taking the e2e-upgrade workflow for example, this runs twice within the
setup-and-test job, along with a bunch of other logic to pull images,
provision a VM, install, upgrade, downgrade and so on. Even with all of
those other steps, it should be possible for two full runs of the CLI
connectivity tests to complete and report failures within the 55m timer.

Signed-off-by: Joe Stringer <joe@cilium.io>