fix(chart): Fix 'loadBalancer' option scope
Technically the 'loadBalancer' option conditionaal here was not properly
closed, meaning that if you somehow omitted this option section from
configuration then Cilium would not apply about half of the possible
option configurations. In practice this was not observed, because the
default values always have a value for the loadBalancer configuration.
Still, it's a logical error so let's fix it to have a valid Helm chart.

Signed-off-by: Matej Líner <matej.liner@lablabs.io>
Signed-off-by: Joe Stringer <joe@cilium.io>